+++ packages/networks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.71](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@0.0.70...@injectivelabs/networks@0.0.71) (2022-02-13)
+
+**Note:** Version bump only for package @injectivelabs/networks
+
+
+
+
+
 ## [0.0.70](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/networks@0.0.68...@injectivelabs/networks@0.0.70) (2022-02-09)