+++ packages/referral-consumer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.11](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/referral-consumer@0.0.10...@injectivelabs/referral-consumer@0.0.11) (2022-02-13)
+
+**Note:** Version bump only for package @injectivelabs/referral-consumer
+
+
+
+
+
 ## [0.0.10](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/referral-consumer@0.0.9...@injectivelabs/referral-consumer@0.0.10) (2022-02-10)