+++ packages/token-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.18](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-utils@0.0.17...@injectivelabs/token-utils@0.0.18) (2022-02-13)
+
+**Note:** Version bump only for package @injectivelabs/token-utils
+
+
+
+
+
 ## [0.0.17](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-utils@0.0.15...@injectivelabs/token-utils@0.0.17) (2022-02-09)