+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.55](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@0.0.54...@injectivelabs/utils@0.0.55) (2022-02-13)
+
+
+### Features
+
+* added number conversion utils ([95d8a09](https://github.com/InjectiveLabs/injective-ts/commit/95d8a090d094421c5700abe34dec5e255da3cc7d))
+
+
+
+
+
 ## [0.0.54](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/utils@0.0.52...@injectivelabs/utils@0.0.54) (2022-02-09)